requirements.txt
aiofiles==0.5.0
click==7.1.2
colorama==0.4.3
fastapi==0.61.0
h11==0.9.0
Pillow==7.2.0
pycryptodome==3.9.8
pydantic==1.6.1
qrcode==6.1
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.6
uvicorn==0.11.8
websockets==8.1
